pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conan-dub"
version = "0.1.0"
description = "Research utility scaffold for the Detective Conan anime dubbing pipeline."
readme = "README.md"
authors = [{ name = "Detective Conan AI Dubbing Team" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pydantic>=1.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cpu = [
    "rich>=13.0",
    "pydub>=0.25",
    "soundfile>=0.12",
    "pandas>=2.0",
    "pyarrow>=16.0",
    "jsonschema>=4.22",
]
gpu = [
    "torch>=2.0",
    "torchaudio>=2.0",
    "transformers>=4.37",
    "whisperx>=3.1",
    "pyannote.audio>=3.1",
    "opencv-python>=4.9",
    "ffmpeg-python>=0.2.0",
    "datasets>=2.19",
    "librosa>=0.10",
]
dev = [
    "black>=24.4",
    "ruff>=0.4",
    "pytest>=7.0",
    "pytest-cov>=4.1",
    "mypy>=1.10",
    "build>=1.2",
]

[project.scripts]
conan-dub = "conan_dub.cli.conan_dub:app"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
show_error_codes = true
